Tests/LibWeb/Text/input/select-text-dialog.html

<!DOCTYPE html>
<script src="include.js"></script>
<dialog id="dialog" open>
    <p id="text">This is a dialog</p>
</dialog>
<script>
    test(() => {
        let { x, y, width, height } = text.getBoundingClientRect();
        x += width / 4;
        y += height / 2;

        internals.mouseDown(x, y);

        x += width / 2;
        internals.mouseMove(x, y);
        internals.mouseUp(x, y);

        println(`"${window.getSelection()}"`);
        printElement(document.activeElement);
    });
</script>
