Tests/LibWeb/Text/expected/select-text-dialog.txt

" is a di"
<DIALOG id="dialog">
